np-41723: moved constants to interface where they are implemented
