--- cristin-organization/src/main/java/no/unit/nva/cristin/organization/common/Constants.java
@@ -1,12 +1,0 @@
-package no.unit.nva.cristin.organization.common;
-
-public class Constants {
-
-    public static final String VERSION = "version";
-    public static final String VERSION_ONE = "1";
-    public static final String VERSION_2023_05_26 = "2023-05-26";
-    public static final String CLIENT_WANTS_VERSION_OF_THE_API_CLIENT = "Client wants version {} of the api client";
-    public static final String CLIENT_DID_NOT_SPECIFY_VERSION_RETURNING_DEFAULT =
-        "Client did not specify version. Returning default version {}";
-
-}
